use shared config for app exp
