+++ packages-exp/performance-exp/rollup.config.release.js
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2019 Google LLC
+ * Copyright 2020 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,35 +15,36 @@
  * limitations under the License.
  */
 
+import json from 'rollup-plugin-json';
 import typescriptPlugin from 'rollup-plugin-typescript2';
 import typescript from 'typescript';
-import json from 'rollup-plugin-json';
-import { importPathTransformer } from '../../scripts/exp/ts-transform-import-path';
-import { es2017BuildsNoPlugin, es5BuildsNoPlugin } from './rollup.shared';
+import pkg from './package.json';
+
+const deps = Object.keys(
+  Object.assign({}, pkg.peerDependencies, pkg.dependencies)
+);
 
 /**
  * ES5 Builds
  */
-const es5BuildPlugins = [
-  typescriptPlugin({
-    typescript,
-    clean: true,
-    transformers: [importPathTransformer]
-  }),
-  json()
-];
+const es5BuildPlugins = [typescriptPlugin({ typescript }), json()];
 
-const es5Builds = es5BuildsNoPlugin.map(build => ({
-  ...build,
-  plugins: es5BuildPlugins,
-  treeshake: {
-    moduleSideEffects: false
+const es5Builds = [
+  {
+    input: 'src/index.ts',
+    output: [
+      { file: pkg.main, format: 'cjs', sourcemap: true },
+      { file: pkg.module, format: 'es', sourcemap: true }
+    ],
+    plugins: es5BuildPlugins,
+    external: id => deps.some(dep => id === dep || id.startsWith(`${dep}/`))
   }
-}));
+];
 
 /**
  * ES2017 Builds
  */
+
 const es2017BuildPlugins = [
   typescriptPlugin({
     typescript,
@@ -51,21 +52,18 @@ const es2017BuildPlugins = [
       compilerOptions: {
         target: 'es2017'
       }
-    },
-    clean: true,
-    transformers: [importPathTransformer]
+    }
   }),
-  json({
-    preferConst: true
-  })
+  json({ preferConst: true })
 ];
 
-const es2017Builds = es2017BuildsNoPlugin.map(build => ({
-  ...build,
-  plugins: es2017BuildPlugins,
-  treeshake: {
-    moduleSideEffects: false
+const es2017Builds = [
+  {
+    input: 'src/index.ts',
+    output: [{ file: pkg.esm2017, format: 'es', sourcemap: true }],
+    plugins: es2017BuildPlugins,
+    external: id => deps.some(dep => id === dep || id.startsWith(`${dep}/`))
   }
-}));
+];
 
 export default [...es5Builds, ...es2017Builds];
